BuyWithAverage algorithm added, TradingAlgorithm.py added
BuyWithAverage algorithm calculates average bid and ask values within a
pre-defined period of ticks and uses this data to decide whether to buy
or to sell.
(Performance with this configuration isn’t that bad: after whole
February +6,56% compared to +2,36% performance of EURUSD)

TradingAlgorithm.py was lost in the repository. Doesn’t work without,
so I added again (supposing deletion was not on purpose). :)