TickBroker now handles orders with numeric order ids, documentation added to Broker
the orders were saved in a list until now. Now they are saved in a dictionary as self._orders[order_id] = order. also, when you make an order, the order_id is returned, so that the TAlgo can use that id to recognize his own orders later. makes 'trader-id' kind of unneccessary. Updates the BuyInTheMorning Algorithm to work with the order id